refactor(docs-retrieval): extract raw-fetch state machine to raw-fetch.ts (Phase 4)
Move the async fetch state machine (Pattern G, HIGH risk) out of
content-fetcher.ts into a dedicated raw-fetch.ts module with explicit
ownership of the drain behavior, trust/HTTPS asymmetry, and error
classification matrix that Phase 4 pinned with the pre-extraction tests.

Symbols moved (atomic, no behavior change):
  - fetchRawHtml
  - tryUrlVariations
  - getContentUrls
  - isJsonContentUrl
  - generateInteractiveLearningVariations
  - enforceHttps          (moved here to avoid a cyclic import: raw-fetch
                           calls it on the direct path; content-fetcher
                           still calls it as the up-front HTTPS gate)
  - FetchError interface
  - FetchRawResult interface

content-fetcher.ts now imports from ./raw-fetch and drops the temporary
exports added in the pre-extraction commit. Public API of docs-retrieval
is unchanged: none of the moved symbols are re-exported from index.ts.

Test plan:
  - raw-fetch.pre-extraction.test.ts -> raw-fetch.test.ts (rename via
    git mv; imports rewritten to ./raw-fetch). 24 tests still pass.
  - npm run typecheck  : clean
  - npm run lint        : clean
  - npm run prettier    : unchanged
  - npm run test:ci     : 2517 passed / 18 skipped / 130 suites

Refs: .refactor/PLAN.md Phase 4 (Pattern G, HIGH)
